Code refactor with useSaveSuggestions.
